od6.0: Added radial basis function gridder to Gridder2D factory
Summary: see D7480

Test Plan: compiles; velocity gridding with triangulation and well log gridders were not working before this patch, and keep on not working. But the RBF gridder used for creating an horizon from a pickset works

Reviewers: raman

Reviewed By: raman

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7758

+++ src/Geometry/explfaultsticksurface.cc
@@ -1244,7 +1244,7 @@ bool ExplFaultStickSurface::setProjTexturePositions( DataPointSet& dps )
 		if ( vs[0]<0 || vs[1]<0 || vs[2]<0 )
 		    continue;
 
-		float w[3];
+		double w[3];
 		interpolateOnTriangle2D( pt, knots[vs[0]], knots[vs[1]],
 			knots[vs[2]], w[0], w[1], w[2] );
 		pos = w[0] * coordlist_->get(knotids[vs[0]]) +
